+++ resource/base/__init__.py
@@ -78,9 +78,9 @@ class Base_Resource(Base_Minimal_Resource):
             req_data_wrap = wrap(req_data)
             if len(req_data_wrap) > 0:
                 for req_data in req_data_wrap:
+                    req_data_lcp = deepcopy(req_data)
+                    req_data_id = req_data.pop('id', id)
                     try:
-                        req_data_lcp = deepcopy(req_data)
-                        req_data_id = req_data.pop('id', id)
                         self.rm_ignore_fields(req_data)
                         obj = self.doc(meta=dict(id=req_data_id),
                                        **req_data)
@@ -96,8 +96,12 @@ class Base_Resource(Base_Minimal_Resource):
                                     resp_data = Unprocessable_Entity_Response(msg)
                                     break
                             resp_data.update(lcp_response=resp_data_lcp)
-                        if not resp_data.error:
+                        force = req_data.get('force', False)
+                        if not resp_data.error or force:
                             obj.save()
+                            if force:
+                                msg = f'Some errors occur but the {self.name} with the id={req_data_id} forcedly created'
+                                resp_data = Unprocessable_Entity_Response(msg)
                         resp_data.add(resp)
                     except Exception as e:
                         self.log.exception(e)
@@ -119,32 +123,36 @@ class Base_Resource(Base_Minimal_Resource):
             req_data_wrap = wrap(req_data)
             if len(req_data_wrap) > 0:
                 for req_data in req_data_wrap:
+                    req_data_lcp = deepcopy(req_data)
+                    req_data_id = req_data.pop('id', id)
                     try:
-                        req_data_lcp = deepcopy(req_data)
-                        req_data_id = req_data.pop('id', id)
                         if len(req_data) == 0:
                             msg = f'Update for {self.name} with id={req_data_id} not necessary'
                             Not_Modified_Response(msg).add(resp)
                         else:
                             self.rm_ignore_fields(req_data)
-                            modified = False
                             obj = self.doc.get(id=req_data_id)
-                            if len(req_data) > 0:
-                                res = obj.update(**req_data)
-                                if res == so.UPDATED:
-                                    modified = True
                             resp_data_lcp = []
                             hndl = self.get_lcp_handler(HTTP_Method.PUT)
                             modified = hndl(instance=obj, req=req_data_lcp,
-                                            resp=resp_data_lcp) or modified
-                            if modified:
-                                msg = f'{self.name.capitalize()} with the id={req_data_id} correctly updated'
-                                resp_data = Ok_Response(msg)
-                            else:
-                                msg = f'{self.name.capitalize()} with the id={req_data_id} no need to update'
-                                resp_data = Not_Modified_Response(msg)
+                                            resp=resp_data_lcp)
+                            resp_data = Ok_Response(f'{self.name.capitalize()} with the id={req_data_id} correctly updated')
                             if len(resp_data_lcp) > 0:
+                                for rdl in resp_data_lcp:
+                                    if rdl['error']:
+                                        resp_data = Unprocessable_Entity_Response(f'Not possible to update a {self.name} with the id={req_data_id}')
+                                        break
                                 resp_data.update(lcp_response=resp_data_lcp)
+                            force = req_data.get('force', False)
+                            if (not resp_data.error or force) and len(req_data) > 0:
+                                res = obj.update(**req_data)
+                                if res == so.UPDATED:
+                                    modified = True
+                                    if force:
+                                        msg = f'Some errors occur but the {self.name} with the id={req_data_id} forcedly updated'
+                                        resp_data = Unprocessable_Entity_Response(msg)
+                            if not modified:
+                                resp_data = Not_Modified_Response(f'{self.name.capitalize()} with the id={req_data_id} no need to update')
                             resp_data.add(resp)
                     except Exception as e:
                         self.log.exception(e)
@@ -184,8 +192,12 @@ class Base_Resource(Base_Minimal_Resource):
                                         resp_data = Unprocessable_Entity_Response(msg)
                                         break
                                 resp_data.update(lcp_response=resp_data_lcp)
-                            if not resp_data.error:
+                            force = req_data.get('force', False)
+                            if not resp_data.error or force:
                                 obj.delete()
+                                if force:
+                                    msg = f'Some errors occur but the {self.name} with the id={hit.meta.id} forcedly deleted'
+                                    resp_data = Unprocessable_Entity_Response(msg)
                             resp_data.add(resp)
                         except Exception as e:
                             self.log.exception(e)
@@ -194,7 +206,8 @@ class Base_Resource(Base_Minimal_Resource):
                                                           exception=e).add(resp)
                 else:
                     msg = f'{self.names.capitalize()} based on the request {{query}} not found'
-                    Not_Found_Response(msg, query=req_data).apply(resp)
+                    Not_Found_Response(msg,
+                                        query=req_data).apply(resp)
             except Exception as e:
                 self.log.exception(e)
                 msg = f'Not possible to delete {self.names} with the request {{query}}'